+++ changelog.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
 ## [0.3.3] Unreleased Github Repo [develop]
+### Added
+- `Hostname` in repr cmd device info
 ## [0.3.2] 2021-10-24
 ### Added
 - mDNS `.local`/`dhcp_hostname` compatibility, so device configuration works across networks