+++ changelog.md
@@ -4,7 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
-## [0.3.3] Unreleased Github Repo [develop]
+## [0.3.4] Unreleased Github Repo [develop]
+
+## [0.3.3] 2021-12-16
+### Added
+- `Host Name` in `__repr__` command device info
 ## [0.3.2] 2021-10-24
 ### Added
 - mDNS `.local`/`dhcp_hostname` compatibility, so device configuration works across networks